+++ .env
@@ -1,1 +1,1 @@
-PYTHONPATH=src
+PYTHONPATH=src/python
