+++ src/python/workbook/a.py
@@ -1,0 +1,3 @@
+
+def blah():
+    print("blah")
