+++ src/python/workbook/b.py
@@ -1,0 +1,6 @@
+from sys import path
+for x in path:
+    print("    ", x)
+import workbook.a
+
+workbook.a.blah()
