+++ src/python/workbook/blah.py
@@ -1,0 +1,3 @@
+from workbook.gen.chapter import Chapter
+from workbook.gen.section import Section
+from workbook.gen.workbook import Workbook
