+++ src/python/workbook/dynamism.py
